Nonroot helper
This helper is introduced to overcome difficulties with SELinux denials
which starts to show up on link phase once file capabilities are used.

This binary needs to have net_bind_service capability.
The launcher is launched with this capability in ambient set.

Futher launcher launches libvird with this capability to
ovecome https://bugzilla.redhat.com/show_bug.cgi?id=1935039.
