+++ pkg/virt-controller/services/template.go
@@ -999,7 +999,11 @@ func (t *templateService) renderLaunchManifest(vmi *v1.VirtualMachineInstance, t
 			"-c",
 			"echo", "bound PVCs"}
 	} else {
-		command = []string{"/usr/bin/virt-launcher",
+		command = []string{"/usr/bin/virt-launcher"}
+		if nonRoot {
+			command = []string{"/usr/bin/virt-launcher-cap", command[0]}
+		}
+		command = append(command, []string{
 			"--qemu-timeout", "5m",
 			"--name", domain,
 			"--uid", string(vmi.UID),
@@ -1012,7 +1016,7 @@ func (t *templateService) renderLaunchManifest(vmi *v1.VirtualMachineInstance, t
 			"--less-pvc-space-toleration", strconv.Itoa(lessPVCSpaceToleration),
 			"--minimum-pvc-reserve-bytes", strconv.FormatUint(reservePVCBytes, 10),
 			"--ovmf-path", ovmfPath,
-		}
+		}...)
 		if nonRoot {
 			command = append(command, "--run-as-nonroot")
 		}
